backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]
mistralai>=1.1.0,<2.0.0
python-multipart
python-dotenv==1.1.0
pydantic
pydantic-settings
PyMuPDF
python-docx
sqlalchemy
passlib[bcrypt]
bcrypt
python-jose[cryptography]
numpy
httpx
